+++ ChangeLog.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- The pathname **WanhiveHosts.IN_MEMORY** for in-memory database.
+
+### Changed
+
+- Setters in the **Message** class use the builder pattern.
+- **Message.MTU** is the default length of the new Messages.
+
+### Removed
+
+- The **Message.freeze** method.
+
 ## [0.4.0] - 2021-01-16
 
 ### Changed